prep-course-portal/requirements.txt
asgiref==3.4.1
attrs==21.2.0
certifi==2021.10.8
charset-normalizer==2.0.9
cycler==0.11.0
dj-database-url==0.5.0
Django==3.1.8
django-environ==0.8.1
django-extensions==3.1.5
django-jsonfield==1.4.1
djangorestframework==3.12.4
drf-spectacular==0.21.0
environ==1.0
fonttools==4.28.3
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
inflection==0.5.1
jsonschema==4.2.1
kiwisolver==1.3.2
matplotlib==3.5.0
numpy==1.21.4
packaging==21.3
pandas==1.3.4
Pillow==8.4.0
psycopg2-binary==2.8
pyparsing==3.0.6
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
requests==2.26.0
setuptools-scm==6.3.2
six==1.16.0
sqlparse==0.4.2
tomli==1.2.2
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
whitenoise==5.3.0
zipp==3.6.0
